+++ liblinphone/tester/bearer-auth.cpp
@@ -669,6 +669,23 @@ static void test_bearer_auth_refresh_token_non_working() {
 	linphone_core_manager_destroy(lcm);
 }
 
+/* this method could be drop when we have a new linphone_core_find_auth_info() variant that can take info account
+ * an auth info type (digest, bearer)
+ **/
+static LinphoneAuthInfo *find_non_bearer_auth_info(LinphoneCore *lc, const char *realm, const char *username) {
+	const bctbx_list_t *ai_list = linphone_core_get_auth_info_list(lc);
+	for (; ai_list != nullptr; ai_list = ai_list->next) {
+		LinphoneAuthInfo *ai = (LinphoneAuthInfo *)ai_list->data;
+		if (linphone_auth_info_get_access_token(ai) != NULL) continue;
+		if (username && linphone_auth_info_get_username(ai) &&
+		    strcmp(linphone_auth_info_get_username(ai), username) == 0 && realm && linphone_auth_info_get_realm(ai) &&
+		    strcmp(linphone_auth_info_get_realm(ai), realm) == 0) {
+			return ai;
+		}
+	}
+	return NULL;
+}
+
 static void provisioning_with_http_bearer_then_register_with_digest_base(bool with_username_in_auth_info) {
 	HttpServerWithBearerAuth httpServer("sip.example.org", "abcdefgh");
 	httpServer.addResource("/provisioning", "application/xml", "rcfiles/marie_digest_auth_xml");
@@ -723,7 +740,7 @@ static void provisioning_with_http_bearer_then_register_with_digest_base(bool wi
 	BC_ASSERT_TRUE(wait_for(lcm->lc, nullptr, &lcm->stat.number_of_LinphoneRegistrationOk, 1));
 	BC_ASSERT_EQUAL(lcm->stat.number_of_authentication_info_requested, 1, int, "%i");
 	const bctbx_list_t *ai_list = linphone_core_get_auth_info_list(lcm->lc);
-	BC_ASSERT_EQUAL((int)bctbx_list_size(ai_list), 3, int, "%i");
+	BC_ASSERT_EQUAL((int)bctbx_list_size(ai_list), 4, int, "%i");
 
 	// ms_message("dump of config:\n %s", linphone_config_dump(linphone_core_get_config(lcm->lc)));
 	/* check that restarting works */
@@ -736,23 +753,39 @@ static void provisioning_with_http_bearer_then_register_with_digest_base(bool wi
 	BC_ASSERT_EQUAL(lcm->stat.number_of_LinphoneConfiguringFailed, 0, int, "%i");
 	BC_ASSERT_EQUAL(lcm->stat.number_of_LinphoneConfiguringSuccessful, 2, int, "%i");
 	BC_ASSERT_TRUE(wait_for(lcm->lc, nullptr, &lcm->stat.number_of_LinphoneRegistrationOk, 2));
-	/* there should still be 3 auth infos */
-	BC_ASSERT_EQUAL((int)bctbx_list_size(linphone_core_get_auth_info_list(lcm->lc)), 3, int, "%i");
+	/* there should still be 4 auth infos */
+	BC_ASSERT_EQUAL((int)bctbx_list_size(linphone_core_get_auth_info_list(lcm->lc)), 4, int, "%i");
 	char *dump = linphone_config_dump(linphone_core_get_config(lcm->lc));
 	ms_message("dump of config after start \n: %s", dump);
 	bctbx_free(dump);
 	account = linphone_core_get_default_account(lcm->lc);
 
+	/* verify that the digest AuthInfo has now ha1 but no more password */
+	const LinphoneAuthInfo *lai = find_non_bearer_auth_info(lcm->lc, "sip.example.org", "marie");
+	if (BC_ASSERT_PTR_NOT_NULL(lai)) {
+		BC_ASSERT_PTR_NULL(linphone_auth_info_get_password(lai));
+		BC_ASSERT_PTR_NOT_NULL(linphone_auth_info_get_ha1(lai));
+	}
+	/* verify that the basic AuthInfo set for hypothetical carddav server was not transformed into ha1 */
+	lai = find_non_bearer_auth_info(lcm->lc, "carddav.example.org", "marie");
+	if (BC_ASSERT_PTR_NOT_NULL(lai)) {
+		BC_ASSERT_PTR_NOT_NULL(linphone_auth_info_get_password(lai));
+		BC_ASSERT_PTR_NULL(linphone_auth_info_get_ha1(lai));
+	}
+
 	/* Test the clearning of the account */
 	linphone_core_remove_account_with_data(lcm->lc, account);
 	BC_ASSERT_TRUE(wait_for(lcm->lc, nullptr, &lcm->stat.number_of_LinphoneRegistrationCleared, 2));
 	if (!with_username_in_auth_info) {
-		/* only the TURN username/pass, that cannot be easily related to the account, should remain .*/
-		BC_ASSERT_EQUAL((int)bctbx_list_size(linphone_core_get_auth_info_list(lcm->lc)), 1, int, "%i");
-	} else {
-		/* The TURN auth info and the bearer one should remain. */
+		/* only the TURN auth info and the bearer one, that cannot be easily related to the account, should remain .*/
 		BC_ASSERT_EQUAL((int)bctbx_list_size(linphone_core_get_auth_info_list(lcm->lc)), 2, int, "%i");
+	} else {
+		/* The TURN auth infos should remain. */
+		BC_ASSERT_EQUAL((int)bctbx_list_size(linphone_core_get_auth_info_list(lcm->lc)), 1, int, "%i");
 	}
+	dump = linphone_config_dump(linphone_core_get_config(lcm->lc));
+	ms_message("dump of config after removing account \n: %s", dump);
+	bctbx_free(dump);
 
 	linphone_core_cbs_unref(cbs);
 	linphone_core_manager_destroy(lcm);